+++ src/components/core/layout/Scrollability/Scrollability-types.ts
@@ -1,0 +1,10 @@
+export interface HorizontalProps {
+  horizontal: boolean;
+}
+
+export interface VerticalProps {
+  vertical: boolean;
+  maxHeight: string;
+}
+
+export type ViewProps = HorizontalProps | VerticalProps;